+++ packages/form/src/layouts/QueryFilter/Actions.tsx
@@ -1,5 +1,6 @@
 import { DownOutlined } from '@ant-design/icons';
-import { IntlType, ProProvider } from '@ant-design/pro-provider';
+import type { IntlType } from '@ant-design/pro-provider';
+import { ProProvider } from '@ant-design/pro-provider';
 import { useIntl } from '@ant-design/pro-provider';
 import { omitBoolean } from '@ant-design/pro-utils';
 import { ConfigProvider, Space } from 'antd';