+++ packages/layout/src/components/AppsLogoComponents/style/default.ts
@@ -23,7 +23,7 @@ const genAppsLogoComponentsDefaultListStyle: GenerateStyle<AppsLogoComponentsTok
           verticalAlign: 'top',
           listStyleType: 'none',
           transition: 'transform 0.2s cubic-bezier(0.333, 0, 0, 1)',
-          borderRadius: token.radiusBase,
+          borderRadius: token.borderRadius,
           '&:hover': {
             backgroundColor: 'rgba(0, 0, 0, 0.03)',
           },