+++ packages/layout/src/components/WaterMark/demos/frontend.tsx
@@ -1,8 +1,6 @@
 /** Title: 前置水印 */
-import { QuestionCircleOutlined } from '@ant-design/icons';
 import type { ProColumns } from '@ant-design/pro-components';
 import { ProTable, TableDropdown, WaterMark } from '@ant-design/pro-components';
-import { Tooltip } from 'antd';
 
 const valueEnum = {
   0: 'close',
@@ -74,21 +72,6 @@ const columns: ProColumns<TableListItem>[] = [
       陈帅帅: { text: '陈帅帅' },
       兼某某: { text: '兼某某' },
     },
-  },
-  {
-    title: (
-      <>
-        创建时间
-        <Tooltip placement="top" title="这是一段描述">
-          <QuestionCircleOutlined style={{ marginInlineStart: 4 }} />
-        </Tooltip>
-      </>
-    ),
-    width: 140,
-    key: 'since',
-    dataIndex: 'createdAt',
-    valueType: 'date',
-    sorter: (a, b) => a.createdAt - b.createdAt,
   },
   {
     title: '备注',