+++ packages/layout/src/style/index.ts
@@ -6,6 +6,11 @@ export interface ProLayoutToken extends ProAliasToken {
   componentCls: string;
 }
 
+const getVersion = () => {
+  if (typeof process === 'undefined') return version;
+  return process?.env?.ANTD_VERSION || version;
+};
+
 /**
  * 主要区别：
  * 需要手动引入 import 'antd/dist/antd.css';
@@ -14,7 +19,7 @@ export interface ProLayoutToken extends ProAliasToken {
  * @returns
  */
 const compatibleStyle: GenerateStyle<ProLayoutToken> = (token) => {
-  if (version.startsWith('5')) {
+  if (getVersion()?.startsWith('5')) {
     return {};
   }
   return {
@@ -64,7 +69,7 @@ const compatibleStyle: GenerateStyle<ProLayoutToken> = (token) => {
             ${token.antCls}-menu-submenu-active, 
             ${token.antCls}-menu-submenu-title:hover`]: {
             color: token?.layout?.sider?.colorTextMenuActive,
-            borderRadius: token.radiusBase,
+            borderRadius: token.borderRadius,
             [`${token.antCls}-menu-submenu-arrow`]: {
               color: token?.layout?.sider?.colorTextMenuActive,
             },
@@ -73,13 +78,13 @@ const compatibleStyle: GenerateStyle<ProLayoutToken> = (token) => {
         [`&${token.antCls}-menu:not(${token.antCls}-menu-horizontal)`]: {
           [`${token.antCls}-menu-item-selected`]: {
             backgroundColor: token?.layout?.sider?.colorBgMenuItemSelected,
-            borderRadius: token.radiusBase,
+            borderRadius: token.borderRadius,
           },
           [`${token.antCls}-menu-item:hover, 
             ${token.antCls}-menu-item-active,
             ${token.antCls}-menu-submenu-title:hover`]: {
             color: token?.layout?.sider?.colorTextMenuActive,
-            borderRadius: token.radiusBase,
+            borderRadius: token.borderRadius,
             backgroundColor: token?.layout?.sider?.colorBgMenuItemHover,
             [`${token.antCls}-menu-submenu-arrow`]: {
               color: token?.layout?.sider?.colorTextMenuActive,
@@ -99,7 +104,7 @@ const compatibleStyle: GenerateStyle<ProLayoutToken> = (token) => {
 
         [`&${token.antCls}-menu-vertical`]: {
           [`${token.antCls}-menu-submenu-selected`]: {
-            borderRadius: token.radiusBase,
+            borderRadius: token.borderRadius,
             color: token?.layout?.sider?.colorTextMenuSelected,
           },
         },
@@ -125,7 +130,7 @@ const compatibleStyle: GenerateStyle<ProLayoutToken> = (token) => {
           ${token.antCls}-menu-item-selected,
           ${token.antCls}-menu-submenu-selected`]: {
             backgroundColor: token?.layout?.header?.colorBgMenuItemSelected,
-            borderRadius: token.radiusBase,
+            borderRadius: token.borderRadius,
             transition: 'none',
             color: token?.layout?.header?.colorTextMenuSelected,
             [`${token.antCls}-menu-submenu-arrow`]: {
@@ -155,7 +160,7 @@ const compatibleStyle: GenerateStyle<ProLayoutToken> = (token) => {
             ${token.antCls}-menu-submenu-active, 
             ${token.antCls}-menu-submenu-title:hover`]: {
             color: token?.layout?.header?.colorTextMenuActive,
-            borderRadius: token.radiusBase,
+            borderRadius: token.borderRadius,
             transition: 'none',
             backgroundColor: token?.layout?.header?.colorBgMenuItemSelected,
             [`${token.antCls}-menu-submenu-arrow`]: {
@@ -165,7 +170,7 @@ const compatibleStyle: GenerateStyle<ProLayoutToken> = (token) => {
 
           [`${token.antCls}-menu-item-selected`]: {
             color: token?.layout?.header?.colorTextMenuSelected,
-            borderRadius: token.radiusBase,
+            borderRadius: token.borderRadius,
             backgroundColor: token?.layout?.header?.colorBgMenuItemSelected,
           },
         },
@@ -195,14 +200,14 @@ const compatibleStyle: GenerateStyle<ProLayoutToken> = (token) => {
       [`${token.antCls}-menu:not(${token.antCls}-menu-horizontal)`]: {
         [`${token.antCls}-menu-item-selected`]: {
           backgroundColor: 'rgba(0, 0, 0, 0.04)',
-          borderRadius: token.radiusBase,
+          borderRadius: token.borderRadius,
           color: token?.layout?.sider?.colorTextMenuSelected,
         },
         [`${token.antCls}-menu-item:hover, 
           ${token.antCls}-menu-item-active,
           ${token.antCls}-menu-submenu-title:hover`]: {
           color: token?.layout?.sider?.colorTextMenuActive,
-          borderRadius: token.radiusBase,
+          borderRadius: token.borderRadius,
           [`${token.antCls}-menu-submenu-arrow`]: {
             color: token?.layout?.sider?.colorTextMenuActive,
           },