+++ packages/table/src/demos/no-option.tsx
@@ -1,7 +1,7 @@
-import { DownOutlined, QuestionCircleOutlined } from '@ant-design/icons';
+import { DownOutlined } from '@ant-design/icons';
 import type { ProColumns } from '@ant-design/pro-components';
-import { ProTable, TableDropdown } from '@ant-design/pro-components';
-import { Button, Tooltip } from 'antd';
+import { ProTable } from '@ant-design/pro-components';
+import { Button } from 'antd';
 
 const valueEnum = {
   0: 'close',
@@ -62,56 +62,14 @@ const columns: ProColumns<TableListItem>[] = [
     },
   },
   {
-    title: '创建者',
-    width: 80,
-    dataIndex: 'creator',
-    valueEnum: {
-      all: { text: '全部' },
-      付小小: { text: '付小小' },
-      曲丽丽: { text: '曲丽丽' },
-      林东东: { text: '林东东' },
-      陈帅帅: { text: '陈帅帅' },
-      兼某某: { text: '兼某某' },
-    },
-  },
-  {
-    title: (
-      <>
-        创建时间
-        <Tooltip placement="top" title="这是一段描述">
-          <QuestionCircleOutlined style={{ marginInlineStart: 4 }} />
-        </Tooltip>
-      </>
-    ),
+    title: '创建时间',
+    tooltip: '这是一段描述',
     width: 140,
     key: 'since',
+    hideInSearch: true,
     dataIndex: 'createdAt',
     valueType: 'date',
     sorter: (a, b) => a.createdAt - b.createdAt,
-  },
-  {
-    title: '备注',
-    dataIndex: 'memo',
-    ellipsis: true,
-    copyable: true,
-  },
-  {
-    title: '操作',
-    width: 180,
-    key: 'option',
-    valueType: 'option',
-    render: () => [
-      <a key="link">链路</a>,
-      <a key="link2">报警</a>,
-      <a key="link3">监控</a>,
-      <TableDropdown
-        key="actionGroup"
-        menus={[
-          { key: 'copy', name: '复制' },
-          { key: 'delete', name: '删除' },
-        ]}
-      />,
-    ],
   },
 ];