+++ packages/table/src/demos/pollinga.tsx
@@ -98,7 +98,7 @@ export default () => {
       }}
       polling={polling || undefined}
       request={async () => {
-        await timeAwait(1000);
+        await timeAwait(2000);
         setTime(Date.now());
         return {
           data: tableListDataSource,