+++ packages/list/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.27](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@2.0.26...@ant-design/pro-list@2.0.27) (2022-11-08)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [2.0.26](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@2.0.25...@ant-design/pro-list@2.0.26) (2022-11-07)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [2.0.25](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@2.0.24...@ant-design/pro-list@2.0.25) (2022-11-03)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
 ## [2.0.24](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@2.0.23...@ant-design/pro-list@2.0.24) (2022-10-28)
 
 **Note:** Version bump only for package @ant-design/pro-list